+++ packages/main/src/editor/converter/src/markdownToHtml.js
@@ -19,27 +19,20 @@
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
  * @link      https://github.com/serlo-org/athene2-assets for the canonical source repository
  */
-import { uploadFolder } from '@serlo/gcloud'
-import * as path from 'path'
-import { Signale } from 'signale'
+import { converter } from '@serlo/markdown'
 
-const bucket = 'assets.serlo.org'
-const source = path.join(__dirname, '..', 'src')
-
-const signale = new Signale({ interactive: true })
-
-run()
-
-async function run() {
-  try {
-    signale.info('Deploying static assets')
-    await uploadFolder({
-      bucket,
-      source,
-      target: 'athene2-assets'
-    })
-    signale.success(`Successfully deployed static assets`)
-  } catch (e) {
-    signale.fatal(e)
-  }
+const renderMarkdown = input => {
+  let html = converter.makeHtml(input)
+  html = html.replace(/"/gm, '"')
+  return html
+    .replace(
+      /<span class="mathInline">%%(.*?)%%<\/span>/gm,
+      '<katexinline>$1</katexinline>'
+    )
+    .replace(
+      /<span class="math">\$\$(.*?)\$\$<\/span>/gm,
+      '<katexblock>$1</katexblock>'
+    )
+    .replace(/\r?\n/gm, '')
 }
+export default renderMarkdown