+++ packages/main/src/editor/helpers.ts
@@ -19,27 +19,24 @@
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
  * @link      https://github.com/serlo-org/athene2-assets for the canonical source repository
  */
-import { uploadFolder } from '@serlo/gcloud'
-import * as path from 'path'
-import { Signale } from 'signale'
+import * as base64 from 'base-64'
+import $ from 'jquery'
+import * as utf8 from 'utf8'
 
-const bucket = 'assets.serlo.org'
-const source = path.join(__dirname, '..', 'src')
+export const parseState = (raw: string): unknown => {
+  const stringifiedState = utf8.decode(base64.decode(raw))
 
-const signale = new Signale({ interactive: true })
-
-run()
-
-async function run() {
-  try {
-    signale.info('Deploying static assets')
-    await uploadFolder({
-      bucket,
-      source,
-      target: 'athene2-assets'
-    })
-    signale.success(`Successfully deployed static assets`)
-  } catch (e) {
-    signale.fatal(e)
+  if (typeof stringifiedState === 'string') {
+    return JSON.parse(stringifiedState)
   }
+
+  return stringifiedState
+}
+
+export const stringifyState = (state: unknown): string => {
+  return base64.encode(utf8.encode(JSON.stringify(state)))
+}
+
+export const getStateFromElement = (element: HTMLElement) => {
+  return parseState($(element).data('rawContent'))
 }