+++ packages/markdown/src/extensions/strike_through.js
@@ -19,27 +19,19 @@
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
  * @link      https://github.com/serlo-org/athene2-assets for the canonical source repository
  */
-import { uploadFolder } from '@serlo/gcloud'
-import * as path from 'path'
-import { Signale } from 'signale'
-
-const bucket = 'assets.serlo.org'
-const source = path.join(__dirname, '..', 'src')
-
-const signale = new Signale({ interactive: true })
-
-run()
-
-async function run() {
-  try {
-    signale.info('Deploying static assets')
-    await uploadFolder({
-      bucket,
-      source,
-      target: 'athene2-assets'
-    })
-    signale.success(`Successfully deployed static assets`)
-  } catch (e) {
-    signale.fatal(e)
-  }
+/* global define */
+var strikethrough = function() {
+  return [
+    {
+      // strike-through
+      // NOTE: showdown already replaced "~" with "~T", so we need to adjust accordingly.
+      type: 'lang',
+      regex: '(~T){2}([^~]+)(~T){2}',
+      replace: function(match, prefix, content, suffix) {
+        return '<del>' + content + '</del>'
+      }
+    }
+  ]
 }
+
+export default strikethrough