+++ packages/markdown/src/index.js
@@ -19,27 +19,32 @@
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
  * @link      https://github.com/serlo-org/athene2-assets for the canonical source repository
  */
-import { uploadFolder } from '@serlo/gcloud'
-import * as path from 'path'
-import { Signale } from 'signale'
+import Showdown from 'showdown'
 
-const bucket = 'assets.serlo.org'
-const source = path.join(__dirname, '..', 'src')
+import codePrepare from './extensions/serlo_code_prepare'
+import injections from './extensions/injections'
+import table from './extensions/table'
+import htmlStrip from './extensions/html_strip'
+import latex from './extensions/latex'
+import atUsername from './extensions/at_username'
+import strikeThrough from './extensions/strike_through'
+import spoiler from './extensions/spoiler'
+import spoilerPrepare from './extensions/spoiler_prepare'
+import latexOutput from './extensions/latex_output'
+import codeOutput from './extensions/serlo_code_output'
 
-const signale = new Signale({ interactive: true })
-
-run()
-
-async function run() {
-  try {
-    signale.info('Deploying static assets')
-    await uploadFolder({
-      bucket,
-      source,
-      target: 'athene2-assets'
-    })
-    signale.success(`Successfully deployed static assets`)
-  } catch (e) {
-    signale.fatal(e)
-  }
-}
+export const converter = new Showdown.Converter({
+  extensions: [
+    codePrepare,
+    injections,
+    table,
+    htmlStrip,
+    latex,
+    atUsername,
+    strikeThrough,
+    spoiler,
+    spoilerPrepare,
+    latexOutput,
+    codeOutput
+  ]
+})